git-ref-format: new refname crate
Introduces a new crate which generalises the refname facilities found in
`git_ext::reference::name`. Namely, the crate provides:

* owned and borrowed types
* more efficient joining
* iterators
* typed "literals"
* structure witnesses
* constants useful for pattern matching

The crate is more conservative about conversions, as it has been found
that those are usually based on conventions which are better handled by
pattern matching (and sometimes are application-specific, such as link
remote tracking branches). Specifically there is no type corresponding
to `OneLevel` from the `git-ext` crate.

Property testing revealed some edge cases in the `check_ref_format`
function, the fixes have, however, **not** been cross-ported to
`git-ext`.

The work here started out as changes to the `git-ext` crate, but were
moved to a separate crate for two reasons:

1. There are no dependencies on git libraries, which makes the crate a
   nicer dependency within the many-crate workspace, or potentially even
   standalone.
2. The changes are somewhat significant, so moving the rest of the
   workspace over from `git-ext` can be done in several patches, which
   also don't need to be combined with this one.

Signed-off-by: Kim Altintop <kim@eagain.st>